renamed styled -> legacy_styled, renamed customStyled -> styled
