+++ docs/src/pages/css-in-js/basics/AdaptingStyledComponents.js
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { styled } from '@material-ui/core/styles';
+import { legacy_styled as styled } from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
 
 const MyButton = styled(({ color, ...other }) => <Button {...other} />)({
