+++ packages/material-ui/src/styled/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './styled';
+export * from './styled';
