+++ packages/material-ui/src/styled/index.js
@@ -1,0 +1,1 @@
+export { default } from './styled';
